CI: test/ion_interoperability: Add tests for the ipn: scheme
This makes the ion interop testing script flexible such that we can
specify the EID scheme to be used. Two additional CI tests are added for
the ipn scheme. Note that these use the newest version of ION while
there still seems to be incompatibility for dtn EIDs with these (see
issue #116).

Signed-off-by: Felix Walter <felix.walter@d3tn.com>