agent_manager, bundle_processor: Support registering RPC-only agents
This adds a second type of agent registration for agents that do not
receive any bundles or events. It is the first step toward splitting all
agent registrations in two parts, depending on the direction of data
flow. In the future, a shared secret will need to be provided to ensure
that an agent is authorized for a given registered agent ID.

Signed-off-by: Felix Walter <felix.walter@d3tn.com>