aap2: Implement Keepalive message support
This sends Keepalive messages if the Client is a subscriber and expects
the reception of Keepalive messages if the Client stays an initiator of
RPC calls. For sent Keepalive messages, a proper acknowledgment is
expected. If no message is received in passive server mode (when uD3TN
is the responder) for twice the configured keepalive timeout, the
connection is closed.

Signed-off-by: Felix Walter <felix.walter@d3tn.com>