CLA, mk: Make SO_REUSEPORT a documented compile-time option
Previously we were setting this every time that it was defined. The
switch from defining _POSIX_C_SOURCE in some files to defining
_XOPEN_SOURCE in the build system made the symbol disappear on Linux
with GCC, thus, making our current AAP test (which reuses ports) fail.

As SO_REUSEPORT is probably unintended in most cases and may rather lead
to unexpected behavior or even security issues, we should disable it by
default. #144

This commit re-introduces _DEFAULT_SOURCE as feature-test macro to
provide the Linux- and BSD-specific API SO_REUSEPORT, which is not
available otherwise.

Fixes: #144

Signed-off-by: Felix Walter <felix.walter@d3tn.com>