nix: Add protobuf for AAP2
To build μD3TN with AAP2, the protobuf language bindings in the
`generated/` directory must be added as an input source.

In addition, `python3Packages.protobuf` and `python3Packages.setuptools`
must be added as a dependency for `python-ud3tn-utils`.

To interactively generating nanopb protobuf C-language bindings via
`make aap2-proto-headers` additional packages are added to the devShells
environment.

Signed-off-by: Maximilian Nitsch <maximilian.nitsch@d3tn.com>