CI: Use Docker runner for NixOS job
This uses the `nixos/nix` Docker image on a standard Docker CI runner to
execute the Nix build job. We already tested and reviewed this in !134 -
it does not greatly increase the pipeline runtime. Another advantage is
that it validates the build without an active Nix cache.

In addition to the configuration in !134 it pins the NixOS image version
to prevent issues based on a different version in this regard (e.g.,
differences in experimental command or submodule support).

Note that currently newer images than `nixos/nix:2.18.1` do not work
because of issues with submodule handling. This is tracked in #160.

Fixes: #158
Signed-off-by: Felix Walter <felix.walter@d3tn.com>