Ensure proper termination on fatal errors in release builds
Previously we were using ASSERT(0|false), which only works in debug
builds. In some cases, this is intended, but we cannot reliably continue
operation in others. We now use abort() where we have to terminate the
program abnormally. In cases where only debug builds should terminate, a
comment is added.

This is done along with the logging updates as the code is typically
related and would cause conflicts otherwise.

Closes: #163
Signed-off-by: Felix Walter <felix.walter@d3tn.com>