CLA: Make infinite retries the default
As we are ensuring that retries happen with a delay in between we can
infinitely retry connections during a contact. This ensures we can
establish a connection even if the peer becomes available significantly
later than the contact starts on our side. The setting can be configured
in `config.mk`.

Signed-off-by: Felix Walter <felix.walter@d3tn.com>