hal_io: Implement log levels for LOG_ERRNO
This enables us to use log levels for `LOG_ERRNO`. The default is the
"warning" level so that we do not need to adjust all calls. The levels
for connection establishment and receiving from sockets are adjusted to
"info" such that common cases like unreachable peers do not
unnecessarily clutter the logs.

Fixes: #172

Signed-off-by: Felix Walter <felix.walter@d3tn.com>